requirements.txt
opencv-python==4.13.0.92
google-genai==1.70.0
pillow==12.2.0
python-dotenv==1.2.2
rich==14.3.3
